continue working on CddExtFlash

+++ Src/Cdd/CddDma/CddDma.h
@@ -6,8 +6,8 @@
 
   void Dma2Stream3SpiTxInit(void);
   void Dma2Stream2SpiRxInit(void);
-  void Dma2Stream2SpiReceive(uint32_t RxData, const size_t DataLen);
-  void Dma2Stream3SpiSend(uint32_t TxData, const size_t DataLen);
+  void Dma2Stream2SpiReceive(uint32_t RxData, const uint32_t DataLen);
+  void Dma2Stream3SpiSend(uint32_t TxData, const uint32_t DataLen);
 
 #endif /* CDD_DMA_2023_08_27_H */
